+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added `TwoElectronIntegralElement`, `MOTwoElectronIntegral` and `TwoElectronIntegrals` for structuring two electron integral information (#243)
 - Add the keyword `dumpactints` to `BlockOutput` (#245).
 - Added fallback keyword argument for `get_final_energy`, `get_gradient`, and `get_structure` that allows to parse these properties from the `.out` file if no JSON output is available (#237).
+- Added `get_frequencies`, `get_imaginary_frequencies`, `is_pes_minimum`, and `is_pes_transition_state` to the `Output` class (#247).
 
 ### Changed
 - Refactored methods from Runner into BaseRunner (#193)